Add active lesson indicator
