+++ src/app/[locale]/page.tsx
@@ -5,7 +5,6 @@ import TimetableForm from "@/views/main-page-view/components/timetable-form/time
 import MainPageView from "@/views/main-page-view/main-page-view";
 
 export default async function Home() {
-    console.log(env.API_BASE_URL);
     const timetableSettings = await getTimetableSettings();
 
     if (!timetableSettings) {
