+++ src/views/main-page-view/components/desktop-timetable/desktop-timetable.tsx
@@ -13,7 +13,7 @@ export default function DesktopTimetable({ timetableSettings }: Props) {
     const { hours, timetableData } = useDesktopTimetable(timetableSettings);
 
     return (
-        <div className="mx-auto flex h-full w-full max-w-[1920px] flex-col p-4">
+        <div className="mx-auto hidden h-full w-full max-w-[1920px] flex-col p-4 lg:flex">
             <DesktopTimetableHeader />
             <DesktopTimetableGrid hours={hours} timetableData={timetableData} />
         </div>
